sqlite versions of the orderBy operator and its variants
